chore: split mrwdefinition into partial impl
